+++ src/queries/events/sessions/index.ts
@@ -3,8 +3,8 @@ export * from "./matches";
 export * from "./translations";
 export * from "./useGetEventSession";
 export * from "./useGetEventSessionAccess";
-export * from "./useGetEventSessionAccessPasses";
 export * from "./useGetEventSessionAccounts";
+export * from "./useGetEventSessionPasses";
 export * from "./useGetEventSessionSpeakers";
 export * from "./useGetEventSessionSponsors";
 export * from "./useGetEventSessionTracks";
